+++ cmake/modules/contrib/NNPack.cmake
@@ -20,7 +20,7 @@ if(USE_NNPACK)
     set(NNPACK_PATH ${CMAKE_CURRENT_SOURCE_DIR}/NNPack)
   endif()
 	set(PTHREAD_POOL_PATH ${NNPACK_PATH}/deps/pthreadpool)
-  file(GLOB NNPACK_CONTRIB_SRC src/contrib/nnpack/*.cc)
+  file(GLOB NNPACK_CONTRIB_SRC src/runtime/contrib/nnpack/*.cc)
   list(APPEND RUNTIME_SRCS ${NNPACK_CONTRIB_SRC})
 	include_directories(${NNPACK_PATH}/include)
 	include_directories(${PTHREAD_POOL_PATH}/include)
