+++ cmake/modules/contrib/Random.cmake
@@ -17,6 +17,6 @@
 
 if(USE_RANDOM)
   message(STATUS "Build with contrib.random")
-  file(GLOB RANDOM_CONTRIB_SRC src/contrib/random/random.cc)
+  file(GLOB RANDOM_CONTRIB_SRC src/runtime/contrib/random/random.cc)
   list(APPEND RUNTIME_SRCS ${RANDOM_CONTRIB_SRC})
 endif(USE_RANDOM)
