+++ cmake/modules/contrib/Sort.cmake
@@ -17,6 +17,6 @@
 
 if(USE_SORT)
   message(STATUS "Build with contrib.sort")
-  file(GLOB SORT_CONTRIB_SRC src/contrib/sort/*.cc)
+  file(GLOB SORT_CONTRIB_SRC src/runtime/contrib/sort/*.cc)
   list(APPEND RUNTIME_SRCS ${SORT_CONTRIB_SRC})
 endif(USE_SORT)
