+++ src/runtime/contrib/cublas/cublas_utils.cc
@@ -6,9 +6,9 @@
  * to you under the Apache License, Version 2.0 (the
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
- * 
+ *
  *   http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -18,27 +18,40 @@
  */
 
 /*!
- *  Copyright (c) 2017 by Contributors
- * \file Use external nnpack library call.
+ * \file Use external cudnn utils function
  */
-#ifndef TVM_CONTRIB_NNPACK_NNPACK_UTILS_H_
-#define TVM_CONTRIB_NNPACK_NNPACK_UTILS_H_
-#include <tvm/runtime/registry.h>
-#include <tvm/runtime/util.h>
+#include "cublas_utils.h"
 #include <dmlc/thread_local.h>
-#include <dmlc/logging.h>
-#include <nnpack.h>
+#include <tvm/runtime/registry.h>
+#include "../../cuda/cuda_common.h"
 
 namespace tvm {
 namespace contrib {
-using namespace runtime;
 
-struct NNPackThreadLocalEntry {
-  pthreadpool_t threadpool{nullptr};
-  static NNPackThreadLocalEntry* ThreadLocal();
-};
 
-bool NNPackConfig(uint64_t nthreads);
+CuBlasThreadEntry::CuBlasThreadEntry() {
+  CHECK_CUBLAS_ERROR(cublasCreate(&handle));
+}
+
+
+CuBlasThreadEntry::~CuBlasThreadEntry() {
+  if (handle) {
+    cublasDestroy(handle);
+    handle = 0;
+  }
+}
+
+
+typedef dmlc::ThreadLocalStore<CuBlasThreadEntry> CuBlasThreadStore;
+
+
+CuBlasThreadEntry* CuBlasThreadEntry::ThreadLocal() {
+  auto stream = runtime::CUDAThreadEntry::ThreadLocal()->stream;
+  CuBlasThreadEntry* retval = CuBlasThreadStore::Get();
+  CHECK_CUBLAS_ERROR(cublasSetStream(retval->handle, static_cast<cudaStream_t>(stream)));
+  return retval;
+}
+
+
 }  // namespace contrib
 }  // namespace tvm
-#endif  // TVM_CONTRIB_NNPACK_NNPACK_UTILS_H_
