+++ src/runtime/contrib/cublas/cublas_utils.h
@@ -6,9 +6,9 @@
  * to you under the Apache License, Version 2.0 (the
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
- * 
+ *
  *   http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -22,8 +22,8 @@
  * \file Use external cudnn utils function
  */
 
-#ifndef TVM_CONTRIB_CUBLAS_CUBLAS_UTILS_H_
-#define TVM_CONTRIB_CUBLAS_CUBLAS_UTILS_H_
+#ifndef TVM_RUNTIME_CONTRIB_CUBLAS_CUBLAS_UTILS_H_
+#define TVM_RUNTIME_CONTRIB_CUBLAS_CUBLAS_UTILS_H_
 
 #include <dmlc/logging.h>
 
@@ -67,4 +67,4 @@ struct CuBlasThreadEntry {
 }  // namespace contrib
 }  // namespace tvm
 
-#endif  // TVM_CONTRIB_CUBLAS_CUBLAS_UTILS_H_
+#endif  // TVM_RUNTIME_CONTRIB_CUBLAS_CUBLAS_UTILS_H_
