+++ src/runtime/contrib/nnpack/nnpack_utils.h
@@ -6,9 +6,9 @@
  * to you under the Apache License, Version 2.0 (the
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
- * 
+ *
  *   http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -18,41 +18,27 @@
  */
 
 /*!
- *  Copyright (c) 2018 by Contributors
- * \file Use external cudnn utils function
+ *  Copyright (c) 2017 by Contributors
+ * \file Use external nnpack library call.
  */
-#include "cublas_utils.h"
-#include <dmlc/thread_local.h>
+#ifndef TVM_RUNTIME_CONTRIB_NNPACK_NNPACK_UTILS_H_
+#define TVM_RUNTIME_CONTRIB_NNPACK_NNPACK_UTILS_H_
 #include <tvm/runtime/registry.h>
-#include "../../runtime/cuda/cuda_common.h"
+#include <tvm/runtime/util.h>
+#include <dmlc/thread_local.h>
+#include <dmlc/logging.h>
+#include <nnpack.h>
 
 namespace tvm {
 namespace contrib {
+using namespace runtime;
 
+struct NNPackThreadLocalEntry {
+  pthreadpool_t threadpool{nullptr};
+  static NNPackThreadLocalEntry* ThreadLocal();
+};
 
-CuBlasThreadEntry::CuBlasThreadEntry() {
-  CHECK_CUBLAS_ERROR(cublasCreate(&handle));
-}
-
-
-CuBlasThreadEntry::~CuBlasThreadEntry() {
-  if (handle) {
-    cublasDestroy(handle);
-    handle = 0;
-  }
-}
-
-
-typedef dmlc::ThreadLocalStore<CuBlasThreadEntry> CuBlasThreadStore;
-
-
-CuBlasThreadEntry* CuBlasThreadEntry::ThreadLocal() {
-  auto stream = runtime::CUDAThreadEntry::ThreadLocal()->stream;
-  CuBlasThreadEntry* retval = CuBlasThreadStore::Get();
-  CHECK_CUBLAS_ERROR(cublasSetStream(retval->handle, static_cast<cudaStream_t>(stream)));
-  return retval;
-}
-
-
+bool NNPackConfig(uint64_t nthreads);
 }  // namespace contrib
 }  // namespace tvm
+#endif  // TVM_RUNTIME_CONTRIB_NNPACK_NNPACK_UTILS_H_
