+++ src/runtime/contrib/random/random.cc
